Add support for multiple entry points
