+++ assets/fontello/css/fontello.css
@@ -1,10 +1,10 @@
 @font-face {
   font-family: 'fontello';
-  src: url('../font/fontello.eot?33620024');
-  src: url('../font/fontello.eot?33620024#iefix') format('embedded-opentype'),
-       url('../font/fontello.woff?33620024') format('woff'),
-       url('../font/fontello.ttf?33620024') format('truetype'),
-       url('../font/fontello.svg?33620024#fontello') format('svg');
+  src: url('../font/fontello.eot?58001839');
+  src: url('../font/fontello.eot?58001839#iefix') format('embedded-opentype'),
+       url('../font/fontello.woff?58001839') format('woff'),
+       url('../font/fontello.ttf?58001839') format('truetype'),
+       url('../font/fontello.svg?58001839#fontello') format('svg');
   font-weight: normal;
   font-style: normal;
 }
@@ -14,49 +14,50 @@
 @media screen and (-webkit-min-device-pixel-ratio:0) {
   @font-face {
     font-family: 'fontello';
-    src: url('../font/fontello.svg?33620024#fontello') format('svg');
+    src: url('../font/fontello.svg?58001839#fontello') format('svg');
   }
 }
 */
- 
+
  [class^="icon-"]:before, [class*=" icon-"]:before {
   font-family: "fontello";
   font-style: normal;
   font-weight: normal;
   speak: none;
- 
+
   display: inline-block;
   text-decoration: inherit;
   width: 1em;
   margin-right: .2em;
   text-align: center;
   /* opacity: .8; */
- 
+
   /* For safety - reset parent styles, that can break glyph codes*/
   font-variant: normal;
   text-transform: none;
- 
+
   /* fix buttons height, for twitter bootstrap */
   line-height: 1em;
- 
+
   /* Animation center compensation - margins should be symmetric */
   /* remove if not needed */
   margin-left: .2em;
- 
+
   /* you can be more comfortable with increased icons size */
   /* font-size: 120%; */
- 
+
   /* Font smoothing. That was taken from TWBS */
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
- 
+
   /* Uncomment for 3D effect */
   /* text-shadow: 1px 1px 1px rgba(127, 127, 127, 0.3); */
 }
- 
+
 .icon-pencil:before { content: '\e800'; } /* '' */
 .icon-cancel:before { content: '\e801'; } /* '' */
 .icon-plus:before { content: '\e802'; } /* '' */
-.icon-ok:before { content: '\e803'; } /* '' */
-.icon-error:before { content: '\e804'; } /* '' */
-.icon-attention:before { content: '\e805'; } /* '' */
+.icon-minus:before { content: '\e803'; } /* '' */
+.icon-ok:before { content: '\e804'; } /* '' */
+.icon-error:before { content: '\e805'; } /* '' */
+.icon-attention:before { content: '\e806'; } /* '' */
