+++ examples/00/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "Tree-shaking"
+	"title": "Dynamic imports"
 }
