+++ examples/00/modules/square.js
