+++ examples/01/example.json
@@ -1,3 +1,4 @@
 {
-	"title": "Default exports"
+	"title": "Multiple Entry Modules",
+    "entryModules": ["main.js", "otherEntry.js"]
 }
