+++ examples/01/modules/hyperCube.js
@@ -1,0 +1,7 @@
+import cube from './cube.js';
+
+// This is only imported by one entry module and
+// shares a chunk with that module
+export default function hyperCube ( x ) {
+	return cube(x) * x;
+}
