+++ examples/01/modules/otherEntry.js
@@ -1,0 +1,4 @@
+// Additional entry modules create new chunks
+import cube from './cube.js';
+
+console.log( cube( 5 ) );
