+++ examples/01/modules/square.js
@@ -1,0 +1,4 @@
+// This is also shared between the entry chunks
+export default function square ( x ) {
+	return x * x;
+}
