+++ examples/02/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "Named exports"
+	"title": "Tree-shaking"
 }
