+++ examples/02/modules/main.js
@@ -1,15 +1,4 @@
-/* NAMED EXPORTS
-   There are many ways to export bindings
-   from an ES2015 module */
-export var foo = 1;
+/* TREE-SHAKING */
+import { cube } from './maths.js';
 
-export function bar () {
-	return foo; // try changing this to `foo++`
-}
-
-function baz () {
-	return bar();
-}
-
-export { baz };
-export * from './qux';
+console.log( cube( 5 ) ); // 125
