+++ examples/03/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "External imports"
+	"title": "Default exports"
 }
