+++ examples/04/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "Static namespaces"
+	"title": "Named exports"
 }
