+++ examples/05/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "Dynamic namespaces"
+	"title": "External imports"
 }
