+++ examples/06/example.json
@@ -1,3 +1,3 @@
 {
-	"title": "Dynamic imports"
+	"title": "Static namespaces"
 }
