+++ examples/06/modules/assert.js
