--- examples/06/modules/maths.js
@@ -1,6 +1,0 @@
-import square from './square.js';
-export {default as square} from './square.js';
-
-export function cube (x ) {
-	return square(x) * x;
-}
