+++ examples/07/example.json
@@ -1,0 +1,3 @@
+{
+	"title": "Dynamic namespaces"
+}
