+++ examples/07/modules/constants.js
