+++ routes/api/examples/_data.js
@@ -7,6 +7,7 @@ function read_json(file) {
 export default fs.readdirSync('examples').filter(file => file[0] !== '.').map(id => {
 	const example = read_json(`examples/${id}/example.json`);
 	example.id = id;
+	example.entryModules = example.entryModules || ['main.js'];
 
 	example.modules = fs
 		.readdirSync(`examples/${id}/modules`)
@@ -15,12 +16,15 @@ export default fs.readdirSync('examples').filter(file => file[0] !== '.').map(id
 			const code = fs
 				.readFileSync(`examples/${id}/modules/${name}`, 'utf-8')
 				.trim();
-			return { name, code };
+			const isEntry = example.entryModules.indexOf(name) >= 0;
+			return { name, code, isEntry };
 		})
 		.sort((a, b) => {
-			return a.name === 'main.js'
-				? -1
-				: b.name === 'main.js' ? 1 : a.name < b.name ? -1 : 1;
+			if (a.name === 'main.js') return -1;
+			if (b.name === 'main.js') return 1;
+			if (a.isEntry) return -1;
+			if (b.isEntry) return 1;
+			return a.name < b.name ? -1 : 1;
 		});
 
 	return example;
